+++ backtest/backtest-engine/src/main/java/com/becker/freelance/backtest/ParameterFilter.java
@@ -2,13 +2,14 @@ package com.becker.freelance.backtest;
 
 import java.util.Map;
 import java.util.function.Predicate;
+import com.becker.freelance.math.Decimal;
 
 public interface ParameterFilter extends AutoCloseable{
 
     static ParameterFilter allOkFilter() {
         return new ParameterFilter() {
             @Override
-            public Predicate<Map<String, Double>> getPredicate() {
+            public Predicate<Map<String, Decimal>> getPredicate() {
                 return map -> true;
             }
 
@@ -19,7 +20,7 @@ public interface ParameterFilter extends AutoCloseable{
         };
     }
 
-    public Predicate<Map<String, Double>> getPredicate();
+    public Predicate<Map<String, Decimal>> getPredicate();
 
     @Override
     void close();
