+++ commons/trading-commons/src/main/java/com/becker/freelance/commons/position/TrailingStopPosition.java
@@ -4,13 +4,14 @@ package com.becker.freelance.commons.position;
 import com.becker.freelance.commons.pair.Pair;
 import com.becker.freelance.commons.signal.Direction;
 import com.becker.freelance.commons.timeseries.TimeSeriesEntry;
+import com.becker.freelance.math.Decimal;
 
 public class TrailingStopPosition extends Position {
 
-    private double trailingStepSize;
+    private Decimal trailingStepSize;
 
-    public TrailingStopPosition(double size, Direction direction, TimeSeriesEntry openPrice, Pair pair,
-                                double stopInPoints, double limitInPoints, double trailingStepSize, double margin) {
+    public TrailingStopPosition(Decimal size, Direction direction, TimeSeriesEntry openPrice, Pair pair,
+                                Decimal stopInPoints, Decimal limitInPoints, Decimal trailingStepSize, Decimal margin) {
         super(size, direction, openPrice, pair, stopInPoints, limitInPoints, PositionType.TRAILING, margin);
         this.trailingStepSize = trailingStepSize;
     }
@@ -21,7 +22,7 @@ public class TrailingStopPosition extends Position {
         throw new UnsupportedOperationException("Not implemented yet");
     }
 
-    public double getTrailingStepSize() {
+    public Decimal getTrailingStepSize() {
         return trailingStepSize;
     }
 }
